pyproject.toml
[project]
name = "luominest"
version = "1.0.0"
description = "LuomiNest AI Agent 桌面应用后端服务"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
full = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
luominest = "src.app:main"

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.idea
  | \.venv
  | venv
  | __pycache__
  | build
  | dist
  | logs
  | data
)/
'''

[tool.flake8]
max-line-length = 120
max-complexity = 10
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
